Move build informations to webhook
